feat(kustomize): support optional services (#6)
* refactor(kustomize): Include shared properties in root kustomization.yml

We need to share commonLabels and namespace among core and optional services, so let's include these fields here instead of in the kustomization-base. Also renames /base to core_services to reflect the kustomization-base refactor.

* feat(kustomize): add support for Kayenta

Users who wish to enable Kayenta will need to manually uncomment the line adding Kayenta in the resources block.

* feat(kustomize): add Fiat as an optional service

Similarly, users will need to enable Fiat by uncommenting this line.

* refactor(style): standardize on .yml

* refactor(kustomization): create spinnaker-secrets

We are not adding any secrets to spinnaker-secrets upstream in the kustomization-base, nor do I think we would ever want to. Instead of merging this secret, let's create it.