feat(core): Add block for overriding versions (#10)
* feat(core): Add block for overriding versions

This commit adds an images: block that sets the version of every
image. To deploy a different spinnaker version, you'd update the
relevant tags. This does currently involve repeating the same
version string a lot, but we will likely improve this in the future.

* fix(core): Use validated version by default

By default let's use master-latest-validated to reduce the chances
of someone testing Spinnaker getting a version with serious bugs.